refactor(analysis): extract Valuator to src/valuation.py with persistable state
