+++ analysis/src/models.py
@@ -7,3 +7,14 @@ class TradeSide:
 @dataclass
 class AverageDraftPosition:
     pass
+
+
+@dataclass
+class PlayerBeliefState:
+    player_id: str
+    guess: float
+    var: float
+    games: float
+    cum_par: float
+    position: str
+    name: str
